Add documentation for format on save
Right now the readme has how to enable and format with xo. This doesn't format on save as mentioned in this issue https://github.com/SamVerschueren/vscode-linter-xo/issues/80. This documentation adds a simple configuration to format on save.

I was thinking about replacing the previous section but wasn't entirely sure that would be the best approach.